Add React/Redux/TypeScript From Create React-App
Add React/Redux/TypeScript from the Create React-App so that the basic
setup is ready to go without having to do extra work to add all of the
packages that are now required to get a frontend working. Some of those
packages include ESLint, prettier, jest, webpack, and babel, to name a
few.

Also, add in Jest watch script to turn on watch mode with a single
commend and speed development.