Add Table Field Component
Add the Table Field component to handle how the orders from the table
will display to the users because the Table component has too much
presentation logic in it as it takes care of the loading part. As a
result, the Table Field component takes care of the field logic,
displaying a suction tag based on the supplier/vendor.