Move Fetch Data Logic to Fetch Data Slice
Move Fetch Data logic to Fetch Data Slice from the component file
because the Slice file should have all of the component's logic while
the component is for displaying the component to the use. Thus, we have
a clean-looking component file while separating the logic from the
component, thus making it look messy.

---

Ref:

https://redux-toolkit.js.org/tutorials/advanced-tutorial#thinking-in-thunks